fixed bug

+++ include/TLattice.h
@@ -36,7 +36,7 @@ public:
   void Evolve();
 
   //! The Lattice is composed by Grid Elements made up with an int (the Index of the Particle if present) and the bool Is_Central
-  typedef mtx::small_vector<uint8_t> GridElement; // type def
+  typedef mtx::small_vector<uint16_t> GridElement; // type def
 
   //Class Setter ang Getter Methods
   //! Put pIndex of the Index-Particle in the pSite.x and pSite.y position of the Lattice
